api_gateway/requirements.txt
google-cloud-logging==3.11.3
pydantic==2.10.4
fastapi==0.115.6
aiohttp==3.11.10
uvicorn==0.34.0
google-auth==2.37.0
requests==2.32.3